ELLE-365 | refactor Complexity components and improve structure
