+++ ui/src/elle/tools/correction/components/CorrectionInfoIcon.jsx
@@ -9,25 +9,20 @@ const alignmentMap = {
   right: 'flex-end'
 };
 
-export default function CorrectionInfoIcon({ children, containerHeight, align = 'right' }) {
+export default function CorrectionInfoIcon({ children, fixedHeight = false, align = 'right', className = '' }) {
   const [modalOpen, setModalOpen] = useState(false);
 
   return (
-    <div style={{ height: containerHeight || 'auto', display: 'flex', justifyContent: alignmentMap[align] }}>
-      <IconButton
-        aria-label="info"
-        color="red"
-        sx
-        onClick={() => setModalOpen(true)}
-      >
+    <div
+      className={className}
+      style={{ height: fixedHeight ? '44px' : 'auto', display: 'flex', justifyContent: alignmentMap[align] }}
+    >
+      <IconButton aria-label="info" color="red" onClick={() => setModalOpen(true)}>
         <InfoIcon className="elle-dark-text" />
       </IconButton>
-      <ModalBase
-        isOpen={modalOpen}
-        setIsOpen={setModalOpen}
-      >
+      <ModalBase isOpen={modalOpen} setIsOpen={setModalOpen}>
         {children}
       </ModalBase>
     </div>
   );
-};
+}
