--- ui/src/elle/tools/correctionV2/CorrectorPage.css
@@ -1,21 +1,0 @@
-.inputField {
-  height: 400px;
-  border: solid 1px lightgray;
-  background-color: white;
-  border-radius: 10px;
-  padding: 0 10px;
-  overflow-y: auto;
-}
-
-*:focus {
-  outline: none;
-}
-
-.tab-context-wrapper {
-  flex: 1;
-}
-
-.correction-tab-group-box {
-  display: flex;
-  justify-content: center;
-}
