--- ui/src/elle/tools/correctionV2/components/CorrectorButton/CorrectorButton.css
@@ -1,4 +1,0 @@
-.corrector-button-container {
-  border-radius: 5px;
-  margin-top: 1rem;
-}
