--- ui/src/elle/tools/correctionV2/components/CorrectrorInput/CorrectorInput.css
@@ -1,28 +1,0 @@
-.ProseMirror {
-  flex: 1;
-  background: #fff;
-  border: 2px solid #a855f7;
-  border-radius: 0.5rem;
-  font-size: 1rem;
-  color: black;
-  text-align: left;
-  height: 400px;
-  overflow-y: auto;
-  padding-top: 2.2rem;
-  padding-left: 0.2rem;
-  white-space: normal !important;
-  word-break: break-word !important;
-}
-
-.corrector-input-icon-bar {
-  position: absolute;
-  top: 2px;
-  left: 2px;
-  width: calc(100% - 4px);
-  padding: 0.2rem;
-  display: flex;
-  align-items: center;
-  justify-content: end;
-  gap: 0.5rem;
-  border-radius: 0.5rem 0.5rem 0 0 !important;
-}
