+++ ui/src/elle/tools/correctionV2/components/editor/CorrectorInput.css
@@ -1,0 +1,43 @@
+.ProseMirror {
+  flex: 1;
+  background: #fff;
+  border: 2px solid #a855f7;
+  border-radius: 0.5rem;
+  font-size: 1rem;
+  color: black;
+  text-align: left;
+  height: 400px;
+  overflow-y: auto;
+  padding-top: 2.2rem;
+  padding-left: 0.2rem;
+  white-space: normal !important;
+  word-break: break-word !important;
+}
+
+.corrector-input-icon-bar {
+  position: absolute;
+  top: 2px;
+  left: 2px;
+  width: calc(100% - 4px);
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 0.25rem;
+  padding: 0.3rem 0.5rem;
+  border-radius: 0.5rem 0.5rem 0 0;
+  border-bottom: 1px solid #e0e0e0;
+  background: rgba(255, 255, 255, 0.95);
+  color: #666;
+  transition: color 0.2s;
+}
+
+.corrector-input-icon-bar:hover {
+  color: #9c27b0;
+}
+
+.corrector-input-icon-bar .container {
+  width: auto;
+  max-width: none;
+  padding: 0;
+  margin: 0;
+}
