--- ui/src/elle/tools/correctionV2/views/CorrectorLayout/CorrectorLayout.css
@@ -1,10 +1,0 @@
-.correctorLayout-container {
-  display: flex;
-  flex-wrap: wrap;
-  padding: 2rem;
-  gap: 2rem;
-
-  @media (max-width: 677px) {
-    padding: 1rem;
-  }
-}
